[feat:reply] Handle UNDO of reply creation/deletion
- change id format of comment replies: use 'cr-...' instead of
'c-reply-...'. Need to do that to avoid conflict with comment ids (which
use 'c-...' as format);

- use MutationObserver to know when a reply was added or removed,
instead of having to call `markCommentsWithReply()` on every possible
action that creates or removes a reply. This is necessary also to be
able to handle UNDO of comment creation/removal, which was not being
handled before;

- generalize `scenesChangedListener` to be a listener of whatever line
type you want to observe. Use this both to listen to scene changes and
to listen to reply changes;

- simplify copy-paste process: do not use any fake comment id, but
instead generate the new comment ids when copying the original content.
Do the same for replies;

- use a more generic way to build HTML content when user selects only
text: create a clone of current selected HTML, which allows other tags
to be copied too; # Please enter the commit message for your changes.
Lines starting