[feat] Enable different CSS styles for each dialog
Add a root class for dialogs according to its target type, so we can
create different CSS styles for them.

Note: if two dialogs use the same target type name, only one of them
will have the class. We need to improve this solution if we want to
differentiate those two dialogs.