tests: Fix accidental treatment of `getCommentId` as synchronous
The `getCommentId()` function uses `helper.waitFor()`, which is
asynchronous. Therefore, `getCommentId()` cannot be used
synchronously. Convert it to an `async` function and await the
resolution wherever it is called.

Also fix the nullish check (`!= null` instead of `!== null`).

I don't understand how any of the tests that used this function ever
worked, because it used to always return `undefined` even though all
callers expect a comment ID. This makes me concerned that the tests
aren't actually testing anything meaningful.