Improve comment edit/delete error handling and tests (#164)
* tests: Use Promises to avoid the pyramid of doom

* tests: Wait for pad creation

* tests: Improve test failure error message

* Don't invoke gritter if there is no error

* Return `no_such_comment` error if comment does not exist

There is no client-side logic to print a pretty localized error
message because this should never happen.

* Change `true` error to `'comment_cannot_be_empty'`

Returning a string gives the user a bit more diagnostic info, and it
is consistent with the other error cases.

* Log erroneous attempts to edit/delete a comment

These are debug and not warning or error so that malicious users can't
DoS the server with log noise.

* Use exceptions for comment edit/delete errors

This is more idiomatic, harder to get wrong, and it surfaces errors
that would otherwise be swallowed.

* Redo client-side error handling for comment edit/delete errors

  * Use async/await instead of callbacks to wait for the response.
  * Use exceptions and try/catch blocks to handle errors closer to
    where they are generated.
  * Let the generic uncaught error handler handle unexpected errors.